code/llm-architecture/base/requirements.txt
torch==2.1.0
datasets==2.14.0
sentencepiece==0.1.99
sacrebleu==2.4.0
numpy==1.26.0
PyYAML==6.0.1
pytest==7.4.0
flake8==6.1.0
matplotlib==3.8.0
jupyter==1.0.0
tensorboard==2.15.0
huggingface_hub==0.20.0
transformers==4.36.2
safetensors==0.4.2
